tests/wpt/web-platform-tests/html/semantics/embedded-content/the-canvas-element/context.unrecognised.unicode.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: context.unrecognised.unicode</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/2dcontext/resources/canvas-tests.css">
<body class="show_output">

<h1>context.unrecognised.unicode</h1>
<p class="desc">Context name which kind of looks like "2d" is unrecognised</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Context name which kind of looks like \"2d\" is unrecognised");
_addTest(function(canvas, ctx) {

_assertSame(canvas.getContext("2\uFF44"), null, "canvas.getContext(\"2\\uFF44\")", "null"); // Fullwidth Latin Small Letter D


});
</script>
